+++ cluster_tools/cluster_tools/__init__.py
@@ -10,6 +10,7 @@ from pathlib import Path
 
 from . import pickling
 from .multiprocessing_logging_handler import get_multiprocessing_logging_setup_fn
+from .schedulers.kube import KubernetesExecutor
 from .schedulers.pbs import PBSExecutor
 from .schedulers.slurm import SlurmExecutor
 from .util import enrich_future_with_uncaught_warning
@@ -326,6 +327,8 @@ def get_executor(environment, **kwargs):
         return SlurmExecutor(**kwargs)
     elif environment == "pbs":
         return PBSExecutor(**kwargs)
+    elif environment == "kubernetes":
+        return KubernetesExecutor(**kwargs)
     elif environment == "multiprocessing":
         global did_start_test_multiprocessing
         if not did_start_test_multiprocessing: